Hoist client and registry modules up to top level
