--- epp-client/src/connection.rs
@@ -1,5 +1,0 @@
-//! Manages registry connections and reading/writing to them
-//! and connects the EppClient instances to them
-
-pub mod client;
-pub mod registry;
